# try to port ASCIIUtility.NarrowUtf16ToAscii to x-plat intrinsics