@adamsitnik 2022-08-10
@tannergooding Since I was not able to get the same perf with `Vector128`, I've added `Vector128` as a fallback when SSE2 and AdvSIMD are not supported which in theory some configs may benefit from.